Move core extensions to jda-ktx module
