+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/api/core/exceptions/InvalidChannelTypeException.kt
@@ -1,10 +1,5 @@
 package io.github.freya022.botcommands.api.core.exceptions
 
-import io.github.freya022.botcommands.api.core.utils.retrieveThreadChannelById
+import dev.freya02.botcommands.jda.ktx.retrieve.InvalidChannelTypeException
 
-/**
- * Exception thrown when retrieving a channel by ID, but the type is incorrect.
- *
- * @see retrieveThreadChannelById
- */
-class InvalidChannelTypeException(message: String) : IllegalArgumentException(message)
+typealias InvalidChannelTypeException = InvalidChannelTypeException
