+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/internal/parameters/resolvers/ChannelResolverFactory.kt
@@ -1,5 +1,8 @@
 package io.github.freya022.botcommands.internal.parameters.resolvers
 
+import dev.freya02.botcommands.jda.ktx.requests.awaitCatching
+import dev.freya02.botcommands.jda.ktx.requests.onErrorResponse
+import dev.freya02.botcommands.jda.ktx.retrieve.retrieveThreadChannelById
 import io.github.freya022.botcommands.api.commands.application.checkGuildOnly
 import io.github.freya022.botcommands.api.commands.application.slash.annotations.ChannelTypes
 import io.github.freya022.botcommands.api.commands.application.slash.options.SlashCommandOption
