+++ BotCommands-jda-ktx/src/main/kotlin/dev/freya02/botcommands/jda/ktx/DeprecatedInBcCore.kt
@@ -1,0 +1,4 @@
+package dev.freya02.botcommands.jda.ktx
+
+@Retention(AnnotationRetention.SOURCE)
+internal annotation class DeprecatedInBcCore
