+++ BotCommands-jda-ktx/src/main/kotlin/dev/freya02/botcommands/jda/ktx/InternalJDAUtils.kt
@@ -1,0 +1,9 @@
+package dev.freya02.botcommands.jda.ktx
+
+import net.dv8tion.jda.api.JDA
+import net.dv8tion.jda.api.requests.RestAction
+import net.dv8tion.jda.api.requests.restaction.CacheRestAction
+import net.dv8tion.jda.internal.requests.DeferredRestAction
+
+internal inline fun <reified R : Any> JDA.deferredRestAction(noinline valueSupplier: () -> R?, noinline actionSupplier: () -> RestAction<R>): CacheRestAction<R> =
+    DeferredRestAction(this, R::class.java, valueSupplier, actionSupplier)
