+++ BotCommands-jda-ktx/src/main/kotlin/dev/freya02/botcommands/jda/ktx/messages/edits/ReplacingWith.kt
@@ -1,0 +1,30 @@
+package dev.freya02.botcommands.jda.ktx.messages.edits
+
+import dev.freya02.botcommands.jda.ktx.DeprecatedInBcCore
+import dev.freya02.botcommands.jda.ktx.messages.toEditData
+import net.dv8tion.jda.api.entities.Message
+import net.dv8tion.jda.api.interactions.InteractionHook
+import net.dv8tion.jda.api.requests.restaction.WebhookMessageEditAction
+import net.dv8tion.jda.api.utils.messages.MessageCreateData
+import net.dv8tion.jda.api.utils.messages.MessageEditData
+
+/**
+ * @see InteractionHook.editOriginal
+ */
+@DeprecatedInBcCore
+fun InteractionHook.replaceWith(data: MessageEditData): WebhookMessageEditAction<Message> =
+    editOriginal(data).setReplace(true)
+
+/**
+ * @see InteractionHook.editOriginal
+ */
+@DeprecatedInBcCore
+fun InteractionHook.replaceWith(data: MessageCreateData): WebhookMessageEditAction<Message> =
+    editOriginal(data.toEditData()).setReplace(true)
+
+/**
+ * @see InteractionHook.editOriginal
+ */
+@DeprecatedInBcCore
+fun InteractionHook.replaceWith(content: String): WebhookMessageEditAction<Message> =
+    editOriginal(content).setReplace(true)
