+++ BotCommands-jda-ktx/src/main/kotlin/dev/freya02/botcommands/jda/ktx/requests/RestResult.kt
@@ -1,8 +1,8 @@
 @file:OptIn(ExperimentalContracts::class)
-@file:Suppress("UNCHECKED_CAST")
 
-package io.github.freya022.botcommands.api.core.utils
+package dev.freya02.botcommands.jda.ktx.requests
 
+import dev.freya02.botcommands.jda.ktx.DeprecatedInBcCore
 import net.dv8tion.jda.api.exceptions.ErrorResponseException
 import net.dv8tion.jda.api.requests.ErrorResponse
 import kotlin.contracts.ExperimentalContracts
@@ -20,6 +20,8 @@ import kotlin.reflect.KClass
  * @see runIgnoringResponseOrNull
  */
 @JvmInline
+@Suppress("UNCHECKED_CAST")
+@DeprecatedInBcCore
 value class RestResult<out T> @PublishedApi internal constructor(
     @PublishedApi internal val value: Any?
 ) {
@@ -172,6 +174,7 @@ value class RestResult<out T> @PublishedApi internal constructor(
  *
  * Returns the original `RestResult` unchanged.
  */
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.onErrorResponseException(block: (ErrorResponseException) -> Unit): RestResult<T> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -190,6 +193,7 @@ inline fun <T> RestResult<T>.onErrorResponseException(block: (ErrorResponseExcep
  * @see ignore
  * @see handle
  */
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.onErrorResponse(block: (ErrorResponse) -> Unit): RestResult<T> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -208,6 +212,7 @@ inline fun <T> RestResult<T>.onErrorResponse(block: (ErrorResponse) -> Unit): Re
  * @see ignore
  * @see handle
  */
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.onErrorResponse(error: ErrorResponse, block: (ErrorResponseException) -> Unit): RestResult<T> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -226,6 +231,7 @@ inline fun <T> RestResult<T>.onErrorResponse(error: ErrorResponse, block: (Error
  *
  * @see handle
  */
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.ignore(predicate: (Throwable) -> Boolean): RestResult<T> {
     contract {
         callsInPlace(predicate, InvocationKind.AT_MOST_ONCE)
@@ -252,6 +258,7 @@ inline fun <T> RestResult<T>.ignore(predicate: (Throwable) -> Boolean): RestResu
  *
  * @see handle
  */
+@DeprecatedInBcCore
 fun <T> RestResult<T>.ignore(vararg responses: ErrorResponse): RestResult<T> =
     ignore { it is ErrorResponseException && it.errorResponse in responses }
 
@@ -266,9 +273,11 @@ fun <T> RestResult<T>.ignore(vararg responses: ErrorResponse): RestResult<T> =
  *
  * @see handle
  */
+@DeprecatedInBcCore
 fun <T> RestResult<T>.ignore(vararg types: KClass<out Throwable>): RestResult<T> =
     ignore { throwable -> types.any { it.isInstance(throwable) } }
 
+@DeprecatedInBcCore
 inline fun <T : R, R> RestResult<T>.recover(predicate: (Throwable) -> Boolean, block: (Throwable) -> R): RestResult<R> {
     contract {
         callsInPlace(predicate, InvocationKind.AT_MOST_ONCE)
@@ -295,6 +304,7 @@ inline fun <T : R, R> RestResult<T>.recover(predicate: (Throwable) -> Boolean, b
  *
  * @see ignore
  */
+@DeprecatedInBcCore
 inline fun <T : R, R> RestResult<T>.recover(vararg responses: ErrorResponse, block: (ErrorResponseException) -> R): RestResult<R> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -306,6 +316,7 @@ inline fun <T : R, R> RestResult<T>.recover(vararg responses: ErrorResponse, blo
     )
 }
 
+@DeprecatedInBcCore
 inline fun <T : R, R> RestResult<T>.recover(vararg types: KClass<out Throwable>, block: (Throwable) -> R): RestResult<R> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -317,6 +328,7 @@ inline fun <T : R, R> RestResult<T>.recover(vararg types: KClass<out Throwable>,
     )
 }
 
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.handle(predicate: (Throwable) -> Boolean, block: (Throwable) -> Unit): RestResult<T> {
     contract {
         callsInPlace(predicate, InvocationKind.UNKNOWN)
@@ -346,6 +358,7 @@ inline fun <T> RestResult<T>.handle(predicate: (Throwable) -> Boolean, block: (T
  *
  * @see ignore
  */
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.handle(vararg responses: ErrorResponse, block: (ErrorResponseException) -> Unit): RestResult<T> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -357,6 +370,7 @@ inline fun <T> RestResult<T>.handle(vararg responses: ErrorResponse, block: (Err
     )
 }
 
+@DeprecatedInBcCore
 inline fun <T> RestResult<T>.handle(vararg types: KClass<out Throwable>, block: (Throwable) -> Unit): RestResult<T> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
@@ -373,6 +387,7 @@ inline fun <T> RestResult<T>.handle(vararg types: KClass<out Throwable>, block: 
  * catching any [Throwable] exception that was thrown from the [block] function execution
  * and encapsulating it as a failure.
  */
+@DeprecatedInBcCore
 inline fun <T> runCatchingRest(block: () -> T): RestResult<T> {
     contract {
         callsInPlace(block, InvocationKind.AT_MOST_ONCE)
