+++ BotCommands-spring/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
 
 dependencies {
     api(projects.botCommandsCore)
+    implementation(projects.botCommandsJdaKtx)
 
     // Logging
     implementation(libs.kotlin.logging)
